% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{original}
\alias{original}
\title{Original data}
\format{
A data frame with 1800 rows and 3 variables:
\describe{
  \item{File}{Sample identifier}
  \item{waveforms}{Waveform scoring}
  \item{time}{waveforms duration}
}
}
\source{
PhD Matteo Ripamonti
}
\usage{
original
}
\description{
A dataset containing the data for 6 Samples
The scoring system is ANA
}
\keyword{datasets}
