% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processed}
\alias{processed}
\title{Compute data}
\format{}
\source{
PhD Matteo Ripamonti
}
\usage{
processed
}
\description{
A dataset containing the processed data
The scoring system is ANA
}
\keyword{datasets}
