% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read data in from a folder}
\usage{
readData(path = ".", estention = c("none", "ANA"))
}
\arguments{
\item{path}{The path to folder containing the data files}

\item{estention}{The files estention identified the type of program produced the output}
}
\description{
Read data in from a folder
}
\examples{
f <- system.file("extdata", package = "Rwaves")
x <- readData(f, estention = "ANA")
}
\author{
Marco Chiapello <chiapello.m@gmail.com>
}
\keyword{IO,}
\keyword{file}
