% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rwaves}
\alias{rwaves}
\title{Compute the parameters}
\usage{
readData(path = ".", estention = c("none", "ANA"))
}
\arguments{
\item{path}{The path to folder containing the data files}

\item{estention}{The files estention identified the type of program produced the output}
}
\description{
Compute the parameters
}
\author{
Marco Chiapello <chiapello.m@gmail.com>
}
\keyword{IO,}
\keyword{file}
