% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processed}
\alias{processed}
\title{Compute data}
\format{A data frame with 6 rows and 14 variables:
\describe{
  \item{File}{Sample identifier}
  \item{data}{Contains the original data}
  \item{f1_1}{total number of 1}
  \item{f2_1}{total duration of 1}
  \item{f3_1}{duration of the 2nd 1 wave}
  \item{f14}{number of probes}
  \item{f24}{total recording time - total duration of 1}
  \item{f2_2}{total duration of 2}
  \item{f2_6}{total duration of 6}
  \item{f1_7}{total number of 7}
  \item{f2_7}{total duration of 7}
  \item{f115_2}{\% probtimeinC 2}
  \item{f115_6}{\% probtimeinC 6}
  \item{f115_7}{\% probtimeinC 7}
}}
\source{
PhD Matteo Ripamonti
}
\usage{
processed
}
\description{
A dataset containing the processed data
}
\keyword{datasets}
