% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rwaves}
\alias{rwaves}
\title{Compute the parameters}
\usage{
rwaves(x)

rwaves(x)
}
\arguments{
\item{x}{it is the object read using the rwaves::readData function}
}
\description{
Compute the parameters

everything between two "1" waves is a probe; it is possible that at the end
of the recording a probe is still ongoing, so there is no "1" wave after
}
\examples{
rwaves(original)
}
\author{
Marco Chiapello <chiapello.m@gmail.com>
}
